+++ src/mod/jedi/relics/StrikeManual.java
@@ -30,7 +30,7 @@ public class StrikeManual
         return this.DESCRIPTIONS[0] + efficiency + this.DESCRIPTIONS[1];
     }
 
-    public int calculateCardDamageRelic(AbstractCard card, AbstractMonster target, int damage)
+    public float calculateCardDamageRelic(AbstractCard card, AbstractMonster target, float damage)
     {
         if (card.hasTag(AbstractCard.CardTags.STRIKE))
         {
@@ -39,7 +39,7 @@ public class StrikeManual
         return damage;
     }
 
-    public int applyPowersRelic(AbstractCard card, int damage)
+    public float applyPowersRelic(AbstractCard card, float damage)
     {
         if (card.hasTag(AbstractCard.CardTags.STRIKE))
         {
